last bug fixes and enemy waves manager added

+++ Eerie_Nights/Assets/Scripts/PowerUps.cs
@@ -25,7 +25,7 @@ public class PowerUps : MonoBehaviour
 
 
 
-    public int Exp;
+    public float Exp;
 
     private void Awake()
     {
